Fixes logic error
The bloom filter was incorrectly identifying items being present or absent
while they were being added to the filter. This stemmed from two errors:

 * Checking if the probed bit was flipped was inverted. That is, previously set
   bits would return true instead of false.
 * Joining each of these individual checks was also inverted.

This would cause the filter to claim that a new item was added only when
the new item matched no bits of any existing item, instead of matching all
bits. This manifested itself particularly when the number of elements inserted
was small, and the elements initialization reflected that. When over
provisioned, the likelihood of hash collision lowered, and the filter appeared
to be more functional.